backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Task Queue
celery==5.3.4
redis==5.0.1

# AI/ML APIs
google-generativeai==0.3.2
google-cloud-speech==2.21.0
openai==1.3.7

# File Processing
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0

# Audio/Video Processing
pydub==0.25.1
moviepy==1.0.3

# OCR
pytesseract==0.3.10

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0 